requirements.txt
asgiref==3.9.1
Django==5.0.9
psycopg2-binary==2.9.10
sqlparse==0.5.3
requests>=2.32.3
django-import-export>=4.1.1
python-decouple==3.8
black==25.1.0
